Use common mocks for seeders
