+++ db/seeders/pending-applications.js
@@ -1,9 +1,38 @@
 'use strict';
-const { prepareSeeds } = require('./data/pending-applications');
+const moment = require('moment');
+const uuidv4 = require('uuid/v4');
+
+const { Users } = require('../models');
+const mockAwardsForAll = require('../../controllers/apply/awards-for-all/mocks');
 
 module.exports = {
-    up: async queryInterface => {
-        const applications = await prepareSeeds();
+    up: async function(queryInterface) {
+        const user = await Users.createUser({
+            username: `${uuidv4()}@example.com`,
+            password: uuidv4(),
+            isActive: true
+        });
+
+        const applications = [
+            moment().add('30', 'days'),
+            moment().add('14', 'days'),
+            moment().add('2', 'days'),
+            moment(),
+            moment().subtract('5', 'days')
+        ].map(function(expiryDate) {
+            return {
+                id: uuidv4(),
+                userId: user.id,
+                formId: 'awards-for-all',
+                applicationData: JSON.stringify(
+                    mockAwardsForAll.mockResponse()
+                ),
+                submissionAttempts: 0,
+                expiresAt: expiryDate.toDate(),
+                createdAt: moment().toDate(),
+                updatedAt: moment().toDate()
+            };
+        });
 
         return queryInterface.bulkInsert(
             'PendingApplications',
@@ -12,7 +41,7 @@ module.exports = {
         );
     },
 
-    down: queryInterface => {
+    down: function(queryInterface) {
         return queryInterface.bulkDelete('PendingApplications', null, {});
     }
 };
